backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==1.10.12
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
supabase==2.3.4
